Segmentation/requirements.txt
numpy~=2.2.1
torch~=2.5.1
pillow~=11.0.0
tqdm~=4.67.1
torchvision~=0.20.1
hydra-core~=1.3.2
omegaconf~=2.3.0
matplotlib~=3.10.0
iopath~=0.1.10
opencv-python~=4.11.0.86